+++ Sources/Client/Model/Channel/Channel+Requests.swift
@@ -124,6 +124,8 @@ public extension Channel {
         Client.shared.markRead(channel: self, completion)
     }
     
+    // MARK: - Events
+    
     /// Send an event.
     /// - Parameters:
     ///   - eventType: an event type.
@@ -131,6 +133,58 @@ public extension Channel {
     @discardableResult
     func send(eventType: EventType, _ completion: @escaping Client.Completion<Event>) -> Cancellable {
         Client.shared.send(eventType: eventType, to: self, completion)
+    }
+    
+    /// Send a keystroke event for the current user.
+    /// - Note: This method should be called from the main thread.
+    /// - Parameter completion: a completion block with `Event`.
+    @discardableResult
+    func keystroke(_ completion: @escaping Client.Completion<Event>) -> Cancellable {
+        keystroke(client: .shared, timerType: DefaultTimer.self, completion)
+    }
+    
+    @discardableResult
+    internal func keystroke(client: Client,
+                            timerType: Timer.Type,
+                            _ completion: @escaping Client.Completion<Event>) -> Cancellable {
+        currentUserTypingTimerControlAtomic.get()?.cancel()
+        
+        let timerControl = timerType.schedule(timeInterval: 15, queue: .main) { [weak self, unowned client] in
+            self?.stopTyping(client: client, completion)
+        }
+        
+        currentUserTypingTimerControlAtomic.set(timerControl)
+        
+        // The user is typing too long, we should resend `.typingStart` event.
+        // The timeout for getting `.typingStop` is 30 sec.
+        if let lastDate = currentUserTypingLastDateAtomic.get(), lastDate.timeIntervalSinceNow > -20 {
+            completion(.success(.typingStart(client.user, cid, .typingStart)))
+            return Subscription.empty
+        }
+        
+        currentUserTypingLastDateAtomic.set(.init())
+        return client.send(eventType: .typingStart, to: self, completion)
+    }
+    
+    /// Send a stop typing event for the current user.
+    /// - Note: This method should be called from the main thread.
+    /// - Parameter completion: a completion block with `Event`.
+    @discardableResult
+    func stopTyping(_ completion: @escaping Client.Completion<Event>) -> Cancellable {
+        stopTyping(client: .shared, completion)
+    }
+    
+    @discardableResult
+    internal func stopTyping(client: Client, _ completion: @escaping Client.Completion<Event>) -> Cancellable {
+        guard currentUserTypingLastDateAtomic.get() != nil else {
+            completion(.success(.typingStop(client.user, cid, .typingStop)))
+            return Subscription.empty
+        }
+        
+        currentUserTypingTimerControlAtomic.get()?.cancel()
+        currentUserTypingLastDateAtomic.set(nil)
+        
+        return client.send(eventType: .typingStop, to: self, completion)
     }
     
     // MARK: - Members